+++ R/AllGenerics.R
@@ -4,12 +4,12 @@
 #' @return 'geneID' return the 'geneID' column of the enriched result which can be converted to data.frame via 'as.data.frame'
 #' @export
 #' @examples
-#' ```r
+#' \dontrun{
 #' data(geneList, package="DOSE")
 #' de <- names(geneList)[1:100]
 #' x <- DOSE::enrichDO(de)
 #' geneID(x)
-#' ```
+#' }
 geneID <- function(x) {
    UseMethod("geneID", x)
 }
@@ -20,12 +20,12 @@ geneID <- function(x) {
 #' @return 'geneInCategory' return a list of genes, by spliting the input gene vector to enriched functional categories
 #' @export
 #' @examples
-#' ```r
+#' \dontrun{
 #' data(geneList, package="DOSE")
 #' de <- names(geneList)[1:100]
 #' x <- DOSE::enrichDO(de)
 #' geneInCategory(x)
-#' ```
+#' }
 geneInCategory <- function(x) {
    UseMethod("geneInCategory", x)
 }